pyproject.toml
[project]
name = "card-fraud-mcp-gateway"
version = "0.1.0"
description = "Enterprise MCP gateway for the Card Fraud platform."
requires-python = ">=3.14"
dependencies = [
    "mcp>=1.0",
    "uvicorn[standard]>=0.30",
    "starlette>=0.40",
    "pydantic-settings>=2.0",
    "asyncpg>=0.30",
    "redis>=5.0",
    "aiokafka>=0.11",
    "aioboto3>=13.0",
    "PyJWT[crypto]>=2.9",
    "httpx>=0.27",
    "structlog>=24.0",
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25",
    "prometheus-client>=0.21",
    "sentry-sdk[fastapi]>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
gateway-dev            = "app.main:dev"
gateway-check          = "app.main:check"
gateway-smoke          = "app.main:smoke"
gateway-export-catalog = "app.main:export_catalog"
doppler-local          = "cli.doppler_local:main"
doppler-test           = "cli.doppler_local:test"
auth0-bootstrap        = "cli.auth0_bootstrap:main"
auth0-verify           = "cli.auth0_verify:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "cli"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
